fix: frontend corrections

+++ app/src/main/java/org/example/auction/AuctionResponse.java
@@ -1,5 +1,6 @@
 package org.example.auction;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record AuctionResponse(
@@ -8,8 +9,8 @@ public record AuctionResponse(
         LocalDateTime startsAt,
         LocalDateTime endsAt,
         String description,
-        Double startPrice,
-        Double currentPrice,
+        BigDecimal startPrice,
+        BigDecimal currentPrice,
         String winnerEmail,
         Integer winnerId,
         Integer ownerId
